Made the script more modular.
Made it so that options get 'parsed' and not always passed on. This
fixes progressbar values passed on to the notification manager when we
just want to display a title.
Made it so that the value of the progressbar doesn't have to be
calculated anymore.
Moved timeout responsibility for sysinfo category to dunstrc